refactor : LadderResult Interface 적용
