+++ src/main/java/nextstep/ladder/domain/result/LadderResult/impl/LadderResultImpl.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain.result;
+package nextstep.ladder.domain.result.LadderResult.impl;
 
 
 import java.util.HashMap;
@@ -7,17 +7,20 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.stream.IntStream;
 import nextstep.ladder.domain.lines.Lines;
+import nextstep.ladder.domain.result.LadderResult.LadderResult;
+import nextstep.ladder.domain.result.Result;
+import nextstep.ladder.domain.result.Results;
 import nextstep.ladder.domain.user.User;
 import nextstep.ladder.domain.user.Users;
 import nextstep.ladder.error.exception.LadderUsersMismatchException;
 import nextstep.ladder.error.exception.NotExistUserException;
 
-public class LadderResult {
+public class LadderResultImpl implements LadderResult {
 
     private final Map<User, Result> ladderResult = new HashMap<>();
     private Results results;
 
-    public LadderResult(Results results, Users users) {
+    public LadderResultImpl(Results results, Users users) {
         if (results.size() != users.size()) {
             throw new LadderUsersMismatchException(results.size(), users.size());
         }
@@ -25,18 +28,21 @@ public class LadderResult {
         this.results = results;
     }
 
+    @Override
     public void calculateLadderResult(Users users, Lines lines) {
         IntStream.range(0, users.size())
             .forEach(userIndex -> ladderResult.put(users.findUser(userIndex),
                 results.findResult(lines.findUserResultIndex(userIndex))));
     }
 
-    public Result findUserDrawResult(User user) throws NotExistUserException {
+    @Override
+    public Result findUserResult(User user) {
         return Optional.ofNullable(ladderResult.get(user))
             .orElseThrow(() -> new NotExistUserException(user.getUserName()));
     }
 
-    public Set<User> keySet() {
+    @Override
+    public Set<User> users() {
         return ladderResult.keySet();
     }
 }
